test: Fix bookings tests
- Changed `prisma.userOnOrg.findMany` to `prisma.userOnOrg.findFirst`

Not necessary to look for all users in an org,
just need to check if the user is in the org.

- Modified return statements in `checkIsUserAdmin`, `checkConflictingBooking`,
and `checkStackedBookings` functions to use strict equality (`!== null`)

This is because we stopped checking .length of the return value (findMany),
and instead just check if the return value is null or not (findFirst).

- Refactored test suite for adding and deleting bookings, also mocking middleware checks

- Updated jest mock in `src/services/test/singleton.ts` to align with named exports

The jest mock should be for `prisma`, not `default` as in the past.

- Added functionality and tests related to IG categories in
`generateRandomIgCategory` function and related generators