feat: add dev login
Currently, the only way to login is to copy the JWT from
the live nusc.club site and paste it into the frontend .env file.

This commit adds a check for dev environment in the /login route
to allow for developer login.

When searching for a user in auth.middleware.ts, the seed script
does not seed the telegramId for Users.
Thus, we look for the user with their username instead.