feat: update submission model and feats
This commit updates the `Submission` model to use an enum for the Semester
and integer for the year.

It also the addSubmission method for the Submission service.
If a unique combination of professor, academic year, semester, and course code
doesn't yet exist in the database (i.e. a `CourseOffering`), we create it.

We use this method for now until we add the CRUD for CourseOffering to more
strictly enforce what CourseOfferings can be created.